requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.9.1
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
comtypes==1.4.9
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dlib @ file:///C:/Users/Ayoub/Downloads/Dlib_Windows_Python3.x-main/Dlib_Windows_Python3.x-main/dlib-19.24.1-cp311-cp311-win_amd64.whl#sha256=6f1a5ee167975d7952b28e0ce4495f1d9a77644761cf5720fb66d7c6188ae496
dnspython==2.7.0
executing==2.0.1
faiss-cpu==1.10.0
fastapi==0.115.7
fastjsonschema==2.19.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.5.4
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
grpcio==1.63.0
gspread==6.2.0
gTTS==2.5.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.28.1
idna==3.7
imutils==0.5.4
ipydatawidgets==4.3.5
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.11
keras==2.15.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.1
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
mtcnn==0.1.1
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.36.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.60.2
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.14.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pillow==10.3.0
platformdirs==4.2.1
plotly==5.23.0
pooch==1.8.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyglet==2.0.17
pygltflib==1.16.2
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
pyparsing==3.1.2
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pythreejs==2.4.2
pyttsx3==2.98
pytz==2024.1
pyvista==0.44.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9.1
safetensors==0.5.2
scikit-learn==1.4.2
scipy==1.13.0
scooby==0.10.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.45.3
streamlit==1.44.1
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.18.0
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.15.2
toml==0.10.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4
tqdm==4.66.5
traitlets==5.14.3
traittypes==0.2.1
trame==3.6.3
trame-client==3.2.5
trame-server==3.1.0
transformers==4.35.2
trimesh==4.4.6
twilio==9.4.4
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.6
ultralytics-thop==2.0.9
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.34.0
vedo==2024.5.2
vtk==9.3.1
watchdog==6.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.11
wrapt==1.16.0
wslink==2.1.2
yarl==1.9.4
